moved typewriter dynamic logic to a separate component
